# [Frontend] speed up import time of vllm.config