by changing some modules in `vllm/multimodal` to lazily import expensive modules like `transformers` or only importing them for type checkers when not used during runtime.

contributes to #14924

I ran on main branch `python -X importtime -c 'import vllm' 2> import.log && tuna import.log`. The visualized call tree shows `vllm.config` accounts for the majority of the total import time at 55.5%.

<img width="1035" alt="image" src="https://github.com/user-attachments/assets/4c2fed6d-9993-4710-98c6-049da9346718" />

On this branch, `vllm.config`'s share decreased to 52.5%.

<img width="1046" alt="image" src="https://github.com/user-attachments/assets/0695cd8f-d8c1-44b6-9886-76bd5349c54f" />

## `python -c 'import vllm'`

on a Google Compute Engine `a2-highgpu-1g` (12 vCPUs, 85 GB Memory) instance with 1 A100 GPU

**~3% decrease** in mean time

### before (main branch commit 94d8ec8d2bcb4ec55e33022b313c7e978edf05e1)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      9.643 s ±  0.140 s    [User: 10.393 s, System: 1.913 s]
  Range (min … max):    9.485 s … 10.084 s    100 runs
```

### after (my PR commit 054f562)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      9.355 s ±  0.096 s    [User: 10.284 s, System: 1.740 s]
  Range (min … max):    9.205 s …  9.711 s    100 runs
```